requirements-dev.txt
coverage
markdown-to-presentation>=0.0.22
pre-commit
pygments-ansi-color
pytest
